Refactor packets in supplier
